fix migrations

+++ database/migrations/2020_09_06_220636_add_userid_to_trades.php
@@ -14,7 +14,7 @@ class AddUseridToTrades extends Migration
     public function up()
     {
         Schema::table('trades', function (Blueprint $table) {
-            $table->unsignedInteger('user_id')->after('id');
+            $table->unsignedBigInteger('user_id')->after('id');
         });
     }
